for #2084, move SetAutoCommit & SetImplicitTransactionsSegment to MySQL & SQLServer module
