+++ sharding-core/sharding-core-parse/sharding-core-parse-mysql/src/main/resources/META-INF/parsing-rule-definition/sharding/mysql/filler-rule-definition.xml
@@ -17,7 +17,6 @@
   -->
 
 <filler-rule-definition>
-    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.dml.predicate.OrPredicateSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.sharding.dml.ShardingOrPredicateFiller" />
     <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.statement.dal.ShowParamSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.dal.ShowParamFiller" />
-    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.antlr.sql.segment.tcl.AutoCommitSegment" filler-class="org.apache.shardingsphere.core.parse.antlr.filler.common.tcl.AutoCommitFiller" />
+    <filler-rule sql-segment-class="org.apache.shardingsphere.core.parse.sql.segment.tcl.AutoCommitSegment" filler-class="org.apache.shardingsphere.core.parse.filler.tcl.MySQLAutoCommitFiller" />
 </filler-rule-definition>
