+++ sharding-core/sharding-core-parse/sharding-core-parse-sqlserver/src/main/resources/META-INF/parsing-rule-definition/sharding/sqlserver/extractor-rule-definition.xml
@@ -19,6 +19,6 @@
 <extractor-rule-definition>
     <extractor-rule id="addColumnDefinition" extractor-class="org.apache.shardingsphere.core.parse.antlr.extractor.impl.ddl.column.AddColumnDefinitionExtractor" />
     <extractor-rule id="modifyColumnDefinition" extractor-class="org.apache.shardingsphere.core.parse.antlr.extractor.impl.ddl.column.ModifyColumnDefinitionExtractor" />
-    <extractor-rule id="addPrimaryKey" extractor-class="org.apache.shardingsphere.core.parse.extractor.SQLServerAddPrimaryKeyExtractor" />
-    <extractor-rule id="setImplicitTransactions" extractor-class="org.apache.shardingsphere.core.parse.antlr.extractor.impl.tcl.SetImplicitTransactionsExtractor" />
+    <extractor-rule id="addPrimaryKey" extractor-class="org.apache.shardingsphere.core.parse.extractor.ddl.SQLServerAddPrimaryKeyExtractor" />
+    <extractor-rule id="setImplicitTransactions" extractor-class="org.apache.shardingsphere.core.parse.extractor.tcl.SetImplicitTransactionsExtractor" />
 </extractor-rule-definition>
